+++ src/RedisEnterpriseCache/RedisEnterpriseCache/ChangeLog.md
@@ -20,6 +20,10 @@
 ## Upcoming Release
 * Preannounced breaking changes. Please refer to https://go.microsoft.com/fwlink/?linkid=2333229
 
+## Version 1.6.0
+* Added a new required property: PublicNetworkAccess for Cluster.
+* Updated the default value of AccessKeysAuthentication property for Database to 'Disabled'.
+
 ## Version 1.5.0
 * Added support for listing all SKUs a cluster can scale to.
 * Added a new enum: NoCluster for Clustering policy.